[OSIS-2973] extracted luy search page common logic in base class
	new file:   base/utils/search.py
	modified:   base/views/learning_units/search/borrowed.py
	modified:   base/views/learning_units/search/common.py
	modified:   base/views/learning_units/search/educational_information.py
	modified:   base/views/learning_units/search/external.py
	modified:   base/views/learning_units/search/proposal.py
	modified:   base/views/learning_units/search/service_course.py
	modified:   base/views/learning_units/search/simple.py
