+++ base/views/learning_units/search/educational_information.py
@@ -23,17 +23,13 @@
 #    see http://www.gnu.org/licenses/.
 #
 ##############################################################################
-from django.contrib.auth.mixins import PermissionRequiredMixin
 from django.utils.translation import ugettext_lazy as _
-from django_filters.views import FilterView
 
 from base.business.learning_units.xls_generator import generate_xls_teaching_material
 from base.forms.learning_unit.search.educational_information import LearningUnitDescriptionFicheFilter
-from base.models.learning_unit_year import LearningUnitYear
-from base.utils.cache import CacheFilterMixin
+from base.utils.search import RenderToExcel
 from base.views.common import remove_from_session
-from base.views.learning_units.search.common import SUMMARY_LIST, RenderToExcel, \
-    SearchMixin
+from base.views.learning_units.search.common import SUMMARY_LIST, BaseLearningUnitSearch
 from learning_unit.api.serializers.learning_unit import LearningUnitDetailedSerializer
 
 
@@ -42,17 +38,10 @@ def _create_xls_teaching_material(view_obj, context, **response_kwargs):
 
 
 @RenderToExcel("xls_teaching_material", _create_xls_teaching_material)
-class LearningUnitDescriptionFicheSearch(PermissionRequiredMixin, CacheFilterMixin, SearchMixin,
-                                         FilterView):
-    model = LearningUnitYear
+class LearningUnitDescriptionFicheSearch(BaseLearningUnitSearch):
     template_name = "learning_unit/search/description_fiche.html"
-    raise_exception = True
-
     filterset_class = LearningUnitDescriptionFicheFilter
     search_type = SUMMARY_LIST
-    permission_required = 'base.can_access_learningunit'
-    cache_exclude_params = 'xls_status',
-
     serializer_class = LearningUnitDetailedSerializer
 
     def get_context_data(self, **kwargs):
@@ -60,10 +49,6 @@ class LearningUnitDescriptionFicheSearch(PermissionRequiredMixin, CacheFilterMix
         self.request.session['ue_search_type'] = str(_('Description fiche status'))
         remove_from_session(self.request, 'search_url')
         context.update({
-            'form': context['filter'].form,
             'is_faculty_manager': self.request.user.person.is_faculty_manager,
-            'search_type': self.search_type,
-            'learning_units_count': context['paginator'].count,
-            'items_per_page': context["paginator"].per_page,
         })
         return context
